+++ src/providers/Auth/AuthProvider.tsx
@@ -5,10 +5,7 @@ import { AuthProviderPropTypes } from "./types";
 import { AuthContext } from "./authContext";
 
 // hooks
-import { useAuthContext } from "./useAuthContext";
-
-// providers
-import { useManager } from "../ManagerProvider";
+import { useManager } from "../useManager";
 
 // lib
 import {
@@ -143,13 +140,4 @@ const AuthProvider = (props: AuthProviderPropTypes) => {
   return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
 };
 
-/**
- * useAuth hook
- * @returns Auth context values from `AuthProvider`.
- */
-const useAuth = () => {
-  return useAuthContext();
-};
-
-// eslint-disable-next-line react-refresh/only-export-components
-export { AuthProvider, useAuth };
+export { AuthProvider };
