+++ src/providers/Auth/index.ts
@@ -1,5 +1,6 @@
 export * from "./types";
 export * from "./AuthProvider";
+export * from "./useAuth";
 export * from "./authContext";
 export * from "./useAuthContext";
 export * from "./useOptionalAuthContext";
