+++ src/providers/Auth/useAuth.ts
@@ -1,0 +1,9 @@
+import { useAuthContext } from "./useAuthContext";
+
+/**
+ * useAuth hook
+ * @returns Auth context values from `AuthProvider`.
+ */
+export const useAuth = () => {
+  return useAuthContext();
+};
