+++ src/providers/ConfigContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+import type { ConfigProviderContextType } from "./types";
+
+export const ConfigContext = createContext<
+  ConfigProviderContextType | undefined
+>(undefined);
