+++ src/providers/DrawerMenuContext.ts
@@ -1,0 +1,14 @@
+import { createContext } from "react";
+
+import type { DrawerMenuContextType } from "./types";
+
+const defaultDrawerMenuContext: DrawerMenuContextType<string> = {
+  addChildItem: () => {},
+  removeChildItem: () => {},
+  clearDynamicItems: () => {},
+  dynamicItems: {},
+};
+
+export const DrawerMenuContext = createContext<DrawerMenuContextType<string>>(
+  defaultDrawerMenuContext,
+);
