+++ src/providers/ManagerContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+import type { ManagerProviderContextType } from "./types";
+
+export const ManagerContext = createContext<
+  ManagerProviderContextType | undefined
+>(undefined);
