+++ src/providers/NotificationContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+import type { NotificationContextType } from "./types";
+
+export const NotificationContext = createContext<
+  NotificationContextType | undefined
+>(undefined);
