+++ src/providers/NotificationProvider.test.tsx
@@ -2,7 +2,8 @@ import { ReactNode } from "react";
 import { act, renderHook } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
 
-import { NotificationProvider, useNotification } from "./NotificationProvider";
+import { NotificationProvider } from "./NotificationProvider";
+import { useNotification } from "./useNotification";
 
 const wrapper = ({ children }: { children: ReactNode }) => (
   <NotificationProvider>{children}</NotificationProvider>
