+++ src/providers/index.ts
@@ -6,4 +6,8 @@ export * from "./Auth";
 export * from "./ConfigProvider";
 export * from "./DrawerMenuProvider";
 export * from "./BottomNavAction";
+export * from "./useConfig";
+export * from "./useDrawerMenu";
+export * from "./useManager";
+export * from "./useNotification";
 export type * from "./types";
